backend/requirements.txt
flask
flask-cors
python-dotenv
anthropic
gunicorn
